refactor/split test
