+++ castagne-backend/Anchor.toml
@@ -15,6 +15,6 @@ cluster = "Localnet"
 wallet = "~/.config/solana/id.json"
 
 [scripts]
-test = "yarn run ts-mocha -p ./tsconfig.json -t 1000000 tests/**/*.ts"
+test = "yarn run ts-mocha -p ./tsconfig.json -t 1000000 tests/castagne/*.ts"
 create-player = "ts-node scripts/create_players.ts"
 read-player = "ts-node scripts/read_players.ts"
